Lab3

+++ code/alexandr-smirnov/src/sort.cpp
@@ -8,9 +8,9 @@ class Sorter::SortImplementation {
     void heapSort(int *Array, int size);
     void mergeSort(int *Array, int size);
     void quickSortProcess(int l, int r);
-    void downHeap(int *mass, long K, int N); 
-    void mergeSortProcess(int *mass, long L, long R);
-    void merge(int *a, int L, int Pivot, long R);
+    void downHeap(int *mass, int K, int N);
+    void mergeSortProcess(int *mass, int L, int R);
+    void merge(int *a, int L, int Pivot, int R);
  private :
     int *elements;
 };
@@ -24,21 +24,21 @@ Sorter::~Sorter(void) {
 
 void Sorter::Sort(int *Array, int size, SortMethod method) {
     switch (method) {
-        case QUICKSORT : 
-        pImpl -> quickSort(Array, size); 
+        case QUICKSORT:
+        pImpl -> quickSort(Array, size);
         break;
 
-        case HEAPSORT: 
-        pImpl -> heapSort(Array, size); 
+        case HEAPSORT:
+        pImpl -> heapSort(Array, size);
         break;
 
-        case MERGESORT: 
-        pImpl -> mergeSort(Array, size); 
+        case MERGESORT:
+        pImpl -> mergeSort(Array, size);
         break;
-    }	
-}    
+    }
+}
 
-void Sorter::SortImplementation::quickSort(int *Array, int size) {	
+void Sorter::SortImplementation::quickSort(int *Array, int size) {
     elements = new int[size];
     elements = Array;
     quickSortProcess(0, size - 1);
@@ -49,9 +49,9 @@ void Sorter::SortImplementation::quickSortProcess(int l, int r) {
     int i = l;
     int j = r;
     while (i <= j) {
-        while (elements[i] < x) 
+        while (elements[i] < x)
             i++;
-        while (elements[j] > x) 
+        while (elements[j] > x)
             j--;
         if (i <= j) {
             int z = elements[i];
@@ -71,24 +71,24 @@ void Sorter::SortImplementation::heapSort(int *Array, int size) {
       elements = new int[size];
       elements = Array;
       int i, tmp;
-      for(i = size / 2 - 1; i >= 0; i--)
+      for (i = size / 2 - 1; i >= 0; i--)
           downHeap(elements, i, size-1);
-      for(i = size - 1; i > 0; i--) {
-          tmp = elements[i]; 
-          elements[i] = elements[0]; 
+      for (i = size - 1; i > 0; i--) {
+          tmp = elements[i];
+          elements[i] = elements[0];
           elements[0] = tmp;
-          downHeap(elements, 0, i - 1); 
+          downHeap(elements, 0, i - 1);
       }
 }
 
-void Sorter::SortImplementation::downHeap(int *mass, long K, int N) {
+void Sorter::SortImplementation::downHeap(int *mass, int K, int N) {
     int new_elem = mass[K], child;
-    while(K <= N / 2) {
+    while (K <= N / 2) {
         child = 2 * K;
-        if (child < N && mass[child] < mass[child + 1]) 
-            child ++;
+        if (child < N && mass[child] < mass[child + 1])
+            child++;
         if (new_elem >= mass[child])
-            break; 
+            break;
         mass[K] = mass[child];
         K = child;
     }
@@ -101,29 +101,29 @@ void Sorter::SortImplementation::mergeSort(int *Array, int size) {
     mergeSortProcess(elements, 0, size - 1);
 }
 
-void Sorter::SortImplementation::mergeSortProcess(int *mass, long L, long R) { 
-    long Pivot;
+void Sorter::SortImplementation::mergeSortProcess(int *mass, int L, int R) {
+    int Pivot;
     if (L < R) {
         Pivot = (L + R) / 2;
-        mergeSortProcess(elements, L, Pivot); 
+        mergeSortProcess(elements, L, Pivot);
         mergeSortProcess(elements, Pivot + 1, R);
         merge(elements, L, Pivot, R);
     }
 }
 
-void Sorter::SortImplementation::merge(int *a, int L, int Pivot, long R) {
-    int pos1 = L, pos2 = Pivot + 1, pos3 = 0;  
+void Sorter::SortImplementation::merge(int *a, int L, int Pivot, int R) {
+    int pos1 = L, pos2 = Pivot + 1, pos3 = 0;
     int* temp = new int[R - L + 1];
-    while(pos1 <= Pivot && pos2 <= R) {
+    while (pos1 <= Pivot && pos2 <= R) {
         if (a[pos1] < a[pos2])
             temp[pos3++] = a[pos1++];
         else
             temp[pos3++] = a[pos2++];
     }
-    while(pos2 <= R)
+    while (pos2 <= R)
         temp[pos3++] = a[pos2++];
-    while( pos1 <= Pivot )
+    while ( pos1 <= Pivot )
         temp[pos3++] = a[pos1++];
-    for( pos3 = 0; pos3 < R - L + 1; pos3++ )
+    for (pos3 = 0; pos3 < R - L + 1; pos3++ )
         a[L + pos3] = temp[pos3];
 }
